Virtual Environment
Working initial release of the virtual environment. This includeds test
code where the left and right mouse buttons can be used to control the
user A and user B positions by clicking on the window. overlaps with the
lures, users and static objects are indicated by the color bars above
and below the display area. User A is at the top, UserB is at the
bottom.